+++ src/main/java/dbdr/security/model/LoginCareworkerArgumentResolver.java
@@ -1,15 +1,14 @@
-package dbdr.security.service;
+package dbdr.security.model;
 
-import dbdr.domain.guardian.entity.Guardian;
-import dbdr.domain.guardian.repository.GuardianRepository;
+import dbdr.domain.careworker.repository.CareworkerRepository;
 import dbdr.global.exception.ApplicationError;
 import dbdr.global.exception.ApplicationException;
-import dbdr.security.LoginGuardian;
+import dbdr.security.LoginCareworker;
 import dbdr.security.dto.BaseUserDetails;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.MethodParameter;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
@@ -18,23 +17,25 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 
 @Component
 @RequiredArgsConstructor
-public class LoginGuardianArgumentResolver implements HandlerMethodArgumentResolver {
+public class LoginCareworkerArgumentResolver implements HandlerMethodArgumentResolver {
 
-
-    private final GuardianRepository guardianRepository;
-
+    private final CareworkerRepository careworkerRepository;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(LoginGuardian.class);
+        return parameter.getParameterAnnotation(LoginCareworker.class) != null &&
+            UserDetails.class.isAssignableFrom(parameter.getParameterType());
     }
 
     @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
+    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
+        NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
+
         BaseUserDetails baseUserDetails = (BaseUserDetails) SecurityContextHolder.getContext()
             .getAuthentication().getPrincipal();
-        return guardianRepository.findById(baseUserDetails.getId())
+        return careworkerRepository.findById(baseUserDetails.getId())
             .orElseThrow(
                 () -> new ApplicationException(ApplicationError.USER_NOT_FOUND));
+
     }
 }
